finally fixed main introduction section  responsiveness for all screen sizes.
image used to take same space in the intro when relatively positioned and leave empty white space behind, and thus refuse to be centered in small screens, it always ends up lopsided.
cropping the picture and centering it manually in the into div win smaller screens made sure it always stays in the center.