Moved images out of link elements since that messed up with element width and position.
when inside an <a> element the images behaved weirdly, the center main image suddenly disappeared for no reason, also the images at the side moved out of screen despite the bounds of their container staying in the same place.
i moved the images out of the element and then positioned it absolutely to always stay on top and get clicked, I also made it responsive.